Split web_server module into separate components
